requirements.txt
google-adk
google-generativeai
arxiv
pdfplumber
pymupdf
markdownify
pydantic
tenacity
fastapi
uvicorn
rich
